refactor: move agent.ts to agent.controller.ts
